Started adding encryption
