+++ tests/SQLiteMetroTests/SQLiteMetroTests.csproj
@@ -121,6 +121,9 @@
     <Compile Include="..\..\src\SQLiteAsync.cs">
       <Link>SQLiteAsync.cs</Link>
     </Compile>
+    <Compile Include="..\..\src\SQLiteAttributes.cs">
+      <Link>SQLiteAttributes.cs</Link>
+    </Compile>
     <Compile Include="..\AsyncTests.cs">
       <Link>AsyncTests.cs</Link>
     </Compile>
